requirements.txt
python_mpd2==0.5.5
requests==2.11.1
